requirements.txt
beautifulsoup4==4.4.1
django-htmlmin==0.8.0
Flask==0.10.1
Flask-Script==2.0.5
Frozen-Flask==0.12
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
Werkzeug==0.10.4
